IPD() should always be returned as uint16 even when being decoded from a uint8
[Tracking bug 24842]

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/python/pbcore/": change = 154170]

+++ pbcore/io/align/_BamSupport.py
@@ -95,7 +95,7 @@ def _makeFramepoints():
         nextOnes = next + grain * np.arange(0, T)
         next = nextOnes[-1] + grain
         framepoints = framepoints + list(nextOnes)
-    return np.array(framepoints, dtype=int)
+    return np.array(framepoints, dtype=np.uint16)
 
 def _makeLookup(framepoints):
     # (frame -> code) involves some kind of rounding
